Sort user table by status
